example/requirements.txt
jinja2
python-multipart
secure
starlette
wtf-tortoise